pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mutation-scatter-plot"
version = "0.3"
description = "Render scatter figures of amino acid or codon mutation frequencies from TSV files using Matplotlib and Bokeh"
readme = "README.md"
license = { text = "CC BY 4.0" }
authors = [
    { name = "Jiří Zahradník" },
    { name = "Martin Mokrejš" },
]
keywords = ["bioinformatics", "mutations", "scatter plot", "codon", "amino acid", "BLOSUM", "matplotlib", "bokeh"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Visualization",
]
requires-python = ">=3.9"
dependencies = [
    "pandas",
    "numpy",
    "matplotlib",
    "biopython",
    "mplcursors",
    "blosum",
    "bokeh",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]
extras = [
    "colorcet",
    "palettable",
]

[project.scripts]
mutation_scatter_plot       = "mutation_scatter_plot.mutation_scatter_plot.cli:main"
calculate_codon_frequencies = "mutation_scatter_plot.calculate_codon_frequencies.cli:main"
count_motifs_in_sequences   = "mutation_scatter_plot.scripts.count_motifs_in_sequences:main"
alignment2dots              = "mutation_scatter_plot.scripts.alignment2dots:main"
split_fasta_entries_by_lengths = "mutation_scatter_plot.scripts.split_fasta_entries_by_lengths:main"

[project.urls]
Homepage = "https://www.biorxiv.org/content/10.1101/2025.04.23.650148v1"

[tool.setuptools.packages.find]
where = ["src"]
